Changes entity attributes to Int32

+++ discount_ascii_warehouse/CDP+Warehouse.swift
@@ -10,12 +10,12 @@ import CoreData
 
 public extension Warehouse {
 
-    @NSManaged var uid: Int16
+    @NSManaged var uid: Int32
     @NSManaged var id: String
     @NSManaged var face: String
     @NSManaged var price: Float
-    @NSManaged var size: Int16
-    @NSManaged var stock: Int16
+    @NSManaged var size: Int32
+    @NSManaged var stock: Int32
     @NSManaged var type: String
     @NSManaged var tag: Set<Tag>
